+++ __tests__/app/page.error.js
@@ -1,0 +1,9 @@
+import { ErrorPage as Base } from '../../src/app';
+
+export default class ErrorPage extends Base {
+
+	componentDidMount() {
+		super.componentDidMount();
+		console.log('custom error');
+	}
+}